==> script/00_00_install_packages.R <==
#script/00_00_install_packages.R
# CRANミラーを非対話でも固定（未設定/@CRAN@ の場合だけ上書き）
if (is.null(getOption("repos")) || getOption("repos")[["CRAN"]] %in% c("@CRAN@", NULL)) {
  options(repos = c(CRAN = "https://cloud.r-project.org"))
}
install.packages("cartogram")
install.packages("estimatr")
install.packages("fixest")
install.packages("GGally")
install.packages("mapview")
install.packages("mlogit")
install.packages("modelsummary")
install.packages("nngeo")
install.packages("openxlsx")
install.packages("osrm")
install.packages("plotly")
install.packages("psych")
install.packages("RColorBrewer")
install.packages("sf")
install.packages("spatialreg")
install.packages("spData")
install.packages("spdep")
install.packages("tidyverse")
install.packages("units")
install.packages("dplyr")
install.packages("stringr")
install.packages("tibble")
install.packages("janitor")
install.packages("purrr")
install.packages("readr")
install.packages("here")
install.packages("languageserver")
